requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.4
decorator==4.4.0
defusedxml==0.6.0
Django==1.11
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==11.1.0
django-bower==5.2.0
django-scheduler==0.8.8
django-tables2==2.1.0
entrypoints==0.3
eventbrite==3.3.3
factory-boy==2.12.0
Faker==2.0.1
flake8==3.7.8
icalendar==4.0.3
idna==2.8
ipdb==0.12.2
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.15.1
mccabe==0.6.1
oauthlib==3.1.0
parso==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
text-unidecode==1.2
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
